requirements.txt
absl-py==2.3.1
altair==5.5.0
asttokens==3.0.0
astunparse==1.6.3
attrs==25.4.0
audioop-lts==0.2.2
audioread==3.1.0
beautifulsoup4==4.14.2
blinker==1.9.0
cachetools==6.2.2
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
comm==0.2.3
contourpy==1.3.3
cycler==0.12.1
debugpy==1.8.17
decorator==5.2.1
executing==2.2.1
filelock==3.20.0
flatbuffers==25.9.23
fonttools==4.60.1
gast==0.6.0
gdown==5.2.0
gitdb==4.0.12
gitpython==3.1.45
google-pasta==0.2.0
grpcio==1.76.0
h5py==3.15.1
idna==3.11
imageio==2.37.2
imbalanced-learn==0.14.0
ipykernel==7.1.0
ipython==9.7.0
ipython-pygments-lexers==1.1.1
jedi==0.19.2
jinja2==3.1.6
joblib==1.5.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter-client==8.6.3
jupyter-core==5.9.1
keras==3.12.0
kiwisolver==1.4.9
lazy-loader==0.4
libclang==18.1.1
librosa==0.11.0
lightgbm==4.6.0
llvmlite==0.45.1
markdown==3.10
markdown-it-py==4.0.0
markupsafe==3.0.3
matplotlib==3.10.7
matplotlib-inline==0.2.1
mdurl==0.1.2
ml-dtypes==0.5.3
msgpack==1.1.2
namex==0.1.0
narwhals==2.11.0
nest-asyncio==1.6.0
networkx==3.5
numba==0.62.1
numpy==2.2.6
opencv-python==4.12.0.88
opt-einsum==3.4.0
optree==0.17.0
packaging==25.0
pandas==2.3.3
parso==0.8.5
pexpect==4.9.0
pillow==12.0.0
platformdirs==4.5.0
pooch==1.8.2
prompt-toolkit==3.0.52
protobuf==6.33.0
psutil==7.1.3
ptyprocess==0.7.0
pure-eval==0.2.3
pyarrow==21.0.0
pycparser==2.23
pydeck==0.9.1
pygments==2.19.2
pyparsing==3.2.5
pysocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2025.2
pyzmq==27.1.0
rawpy==0.25.1
referencing==0.37.0
requests==2.32.5
rich==14.2.0
rpds-py==0.28.0
scikit-image==0.25.2
scikit-learn==1.7.2
scipy==1.16.3
seaborn==0.13.2
setuptools==80.9.0
six==1.17.0
smmap==5.0.2
smote==0.1
soundfile==0.13.1
soupsieve==2.8
soxr==1.0.0
stack-data==0.6.3
standard-aifc==3.13.0
standard-chunk==3.13.0
standard-sunau==3.13.0
streamlit==1.51.0
tenacity==9.1.2
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
termcolor==3.2.0
threadpoolctl==3.6.0
tifffile==2025.10.16
toml==0.10.2
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
typing-extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
watchdog==6.0.0
wcwidth==0.2.14
werkzeug==3.1.3
wheel==0.45.1
wrapt==2.0.1
